+++ mycli/main.py
@@ -1425,10 +1425,13 @@ class MyCli:
 @click.option("--ssh-key-filename", help="Private key filename (identify file) for the ssh connection.")
 @click.option("--ssh-config-path", help="Path to ssh configuration.", default=os.path.expanduser("~") + "/.ssh/config")
 @click.option("--ssh-config-host", help="Host to connect to ssh server reading from ssh configuration.")
-@click.option("--ssl-mode", "ssl_mode", default="auto", help="Set desired SSL behavior. auto=preferred, on=required, off=off.", type=str)
 @click.option(
-    "--ssl/--no-ssl", "ssl_enable", is_flag=True, default=True, help="Enable SSL for connection (automatically enabled with other flags)."
+    "--ssl-mode",
+    "ssl_mode",
+    help="Set desired SSL behavior. auto=preferred, on=required, off=off.",
+    type=click.Choice(["auto", "on", "off"]),
 )
+@click.option("--ssl/--no-ssl", "ssl_enable", default=None, help="Enable SSL for connection (automatically enabled with other flags).")
 @click.option("--ssl-ca", help="CA file in PEM format.", type=click.Path(exists=True))
 @click.option("--ssl-capath", help="CA directory.")
 @click.option("--ssl-cert", help="X509 cert in PEM format.", type=click.Path(exists=True))
@@ -1444,8 +1447,6 @@ class MyCli:
     is_flag=True,
     help=("""Verify server's "Common Name" in its cert against hostname used when connecting. This option is disabled by default."""),
 )
-# as of 2016-02-15 revocation list is not supported by underling PyMySQL
-# library (--ssl-crl and --ssl-crlpath options in vanilla mysql client)
 @click.version_option(__version__, "-V", "--version", help="Output mycli's version.")
 @click.option("-v", "--verbose", is_flag=True, help="Verbose output.")
 @click.option("-D", "--database", "dbname", help="Database to use.")
@@ -1541,6 +1542,15 @@ def cli(
         warn=warn,
         myclirc=myclirc,
     )
+
+    if ssl_enable is not None:
+        click.secho(
+            "Warning: The --ssl/--no-ssl CLI options will be deprecated in a future release. "
+            "Please use the ssl_mode config or --ssl-mode CLI options instead.",
+            err=True,
+            fg="yellow",
+        )
+
     if list_dsn:
         try:
             alias_dsn = mycli.config["alias_dsn"]
@@ -1637,13 +1647,21 @@ def cli(
             ssl_verify_server_cert = ssl_verify_server_cert or (params[0].lower() == 'true')
             ssl_enable = True
 
-    if ssl_mode not in ("auto", "on", "off"):
-        click.secho(f"Invalid option provided for --ssl-mode: {ssl_mode}. See --help for valid options.", err=True, fg="red")
-        sys.exit(1)
-
     ssl_mode = ssl_mode or mycli.ssl_mode  # cli option or config option
 
-    if ssl_mode in ("auto", "on"):
+    # if there is a mismatch between the ssl_mode value and other sources of ssl config, show a warning
+    # specifically using "is False" to not pickup the case where ssl_enable is None (not set by the user)
+    if ssl_enable and ssl_mode == "off" or ssl_enable is False and ssl_mode in ("auto", "on"):
+        click.secho(
+            f"Warning: The current ssl_mode value of '{ssl_mode}' is overriding the value provided by "
+            f"either the --ssl/--no-ssl CLI options or a DSN URI parameter (ssl={ssl_enable}).",
+            err=True,
+            fg="yellow",
+        )
+
+    # configure SSL if ssl_mode is auto/on or if
+    # ssl_enable = True (from --ssl or a DSN URI) and ssl_mode is None
+    if ssl_mode in ("auto", "on") or (ssl_enable and ssl_mode is None):
         ssl = {
             "mode": ssl_mode,
             "enable": ssl_enable,
@@ -1655,7 +1673,6 @@ def cli(
             "tls_version": tls_version,
             "check_hostname": ssl_verify_server_cert,
         }
-
         # remove empty ssl options
         ssl = {k: v for k, v in ssl.items() if v is not None}
     else: